backend/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
google-generativeai>=0.8.4
python-multipart==0.0.12
python-dotenv==1.1.0
pytest==8.3.3
httpx==0.27.2
pytest-cov==5.0.0